Trait tinyraytracer::blocks::RayCollision

pub trait RayCollision {
    fn ray_intersect(&self, ray: &Ray) -> HitPoint;

    fn collision_normal(&self, hit_point: Vec3) -> Vec3;

    fn collision_material(&self, hit_point: Vec3) -> Material;
}

We need to determine if a ray of light hits a specific object or not. This trait contains the logic of how to determine that.

Required methods

fn ray_intersect(&self, ray: &Ray) -> HitPoint

fn collision_normal(&self, hit_point: Vec3) -> Vec3

fn collision_material(&self, hit_point: Vec3) -> Material

Loading content...